requirements.txt
# Core Framework
streamlit==1.31.1
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0

# LLM & AI
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.3.2  # FREE tier available!
langchain==0.1.9
# langchain-openai==0.0.8  # Removed (requires tiktoken)
# tiktoken==0.6.0  # Removed (requires Rust)

# Vector Database
chromadb==0.4.22
sentence-transformers==2.3.1

# Code execution and safety
docker==7.0.0
RestrictedPython==7.4

# Data processing
pandas>=2.0.0
numpy>=1.26.0

# Database
sqlalchemy==2.0.27

# Server (optional, for API)
fastapi==0.109.2
uvicorn==0.27.1
httpx==0.26.0

# Testing
pytest==8.0.1
pytest-asyncio==0.23.5

# Code quality
black==24.2.0
mypy==1.8.0

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1
